Foreground and background persistence working and logging out if time crossed
